requirements.txt
discord.py==2.3.2
fastapi==0.110.0
uvicorn==0.27.1
jinja2==3.1.3
aiofiles==23.2.1
